+++ src/main/java/roomescape/web/ReservationController.java
@@ -1,6 +1,5 @@
-package roomescape;
+package roomescape.web;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicLong;
 import org.springframework.http.ResponseEntity;
@@ -11,19 +10,23 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import roomescape.dto.ReservationFindAllResponse;
-import roomescape.dto.ReservationFindResponse;
-import roomescape.dto.ReservationSaveRequest;
+import roomescape.dao.Reservation;
+import roomescape.dao.ReservationDao;
+import roomescape.dao.ReservationMemoryDao;
+import roomescape.web.dto.ReservationFindAllResponse;
+import roomescape.web.dto.ReservationFindResponse;
+import roomescape.web.dto.ReservationSaveRequest;
 
 @RequestMapping("/reservations")
 @Controller
 public class ReservationController {
 
+    private final ReservationDao reservationDao = new ReservationMemoryDao();
     private final AtomicLong counter = new AtomicLong();
-    private final List<Reservation> reservations = new ArrayList<>();
 
     @GetMapping
     public ResponseEntity<List<ReservationFindAllResponse>> findAllReservation() {
+        List<Reservation> reservations = reservationDao.findAll();
         List<ReservationFindAllResponse> response = reservations.stream()
                 .map(ReservationFindAllResponse::from)
                 .toList();
@@ -33,7 +36,7 @@ public class ReservationController {
     @PostMapping
     public ResponseEntity<ReservationFindResponse> saveReservation(@RequestBody ReservationSaveRequest request) {
         Reservation reservation = request.toEntity(counter.incrementAndGet());
-        reservations.add(reservation);
+        reservationDao.save(reservation);
         ReservationFindResponse response = ReservationFindResponse.from(reservation);
         return ResponseEntity.ok(response);
     }
@@ -41,14 +44,12 @@ public class ReservationController {
     @DeleteMapping("/{reservation_id}")
     public ResponseEntity<Void> deleteReservation(@PathVariable(value = "reservation_id") Long id) {
         Reservation reservation = findReservationById(id);
-        reservations.remove(reservation);
+        reservationDao.delete(reservation);
         return ResponseEntity.ok().build();
     }
 
     private Reservation findReservationById(Long id) {
-        return reservations.stream()
-                .filter(reservation -> reservation.getId().equals(id))
-                .findFirst()
+        return reservationDao.findById(id)
                 .orElseThrow();
     }
 }
